feat: Add unit type to Location model and related components
- Updated LocationController to handle unitType in DTOs.
- Modified LocationDTO to include unitType with default value.
- Changed LocationOccupancyDTO to use double for capacity fields.
- Enhanced Location model to include unitType with default value.
- Adjusted LocationThreshold to use double for threshold fields.
- Updated LocationService to manage unitType during updates.
- Modified EditItemModal to retain original quantity and handle unit changes.
- Auto-set unit based on selected location in AddComponentsAndProducts.
- Implemented sorting functionality in LastOperations and Summary components.
- Updated InventoryStatus report to display occupancy percentage with two decimal places.
- Enhanced Location settings page to include unit type selection.
- Updated locationService to include unitType in Location interface.

+++ backend/src/main/java/com/kozimor/wms/Database/Model/DTO/LocationOccupancyDTO.java
@@ -14,9 +14,9 @@ public class LocationOccupancyDTO {
     private String locationCode;
     private String locationName;
     private String locationDescription;
-    private int maxCapacity;
-    private int minThreshold;
-    private int currentOccupancy;
+    private double maxCapacity;
+    private double minThreshold;
+    private double currentOccupancy;
     private double occupancyPercentage; // 0-100
     private int itemCount;
     private boolean isAboveThreshold; // Czy powyżej minimalnego progu
